refactor: move hooks references to index.ts
