+++ public/hooks/index.ts
@@ -4,3 +4,5 @@
  */
 
 export { useSaveChat } from './use_save_chat';
+export { useChatState, ChatStateProvider } from './use_chat_state';
+export { useChatActions } from './use_chat_actions';
