+++ public/tabs/chat/chat_page.tsx
@@ -6,12 +6,11 @@
 import { EuiFlyoutBody, EuiFlyoutFooter, EuiPage, EuiPageBody, EuiSpacer } from '@elastic/eui';
 import React, { useCallback, useState } from 'react';
 import cs from 'classnames';
-import { useChatContext } from '../../contexts/chat_context';
-import { useChatState } from '../../hooks/use_chat_state';
+import { useObservable } from 'react-use';
+import { useChatContext, useCore } from '../../contexts';
+import { useChatState } from '../../hooks';
 import { ChatPageContent } from './chat_page_content';
 import { ChatInputControls } from './controls/chat_input_controls';
-import { useObservable } from 'react-use';
-import { useCore } from '../../contexts/core_context';
 
 interface ChatPageProps {
   className?: string;
