+++ public/tabs/history/chat_history_page.tsx
@@ -20,8 +20,7 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { FormattedMessage } from '@osd/i18n/react';
 import { useDebounce, useObservable } from 'react-use';
 import cs from 'classnames';
-import { useChatActions } from '../../hooks/use_chat_actions';
-import { useChatState } from '../../hooks/use_chat_state';
+import { useChatActions, useChatState } from '../../hooks';
 import { useChatContext, useCore } from '../../contexts';
 import { ChatHistorySearchList } from './chat_history_search_list';
